+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/domain/project/exporter/ExportOptions.kt
@@ -1,0 +1,3 @@
+package org.wycliffeassociates.otter.common.domain.project.exporter
+
+data class ExportOptions(val chapters: List<Int>)
